+++ ChangeLog.md
@@ -9,6 +9,8 @@ SPDX-License-Identifier: MIT
 
 ## UNRELEASED
 
+* capywfa: SBOM properties for workflow control renamed from `distroclearing:*` to `capywfa:*`
+  Additionally, our property Sw360SourceFileChecked was renamed to Sw360SourceFileCheck
 * lst_to_sbom: add support for guessing Alpine Linux 3.22
 * lst_to_sbom: create valid CycloneDX BOMs using the CycloneDX Python library
 * update requests to 2.32.4 to fix CVE-2024-47081